限时购库表修改

+++ mall-mbg/src/main/java/com/macro/mall/mapper/SmsFlashPromotionMapper.java
@@ -10,7 +10,7 @@ public interface SmsFlashPromotionMapper {
 
     int deleteByExample(SmsFlashPromotionExample example);
 
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(SmsFlashPromotion record);
 
@@ -18,7 +18,7 @@ public interface SmsFlashPromotionMapper {
 
     List<SmsFlashPromotion> selectByExample(SmsFlashPromotionExample example);
 
-    SmsFlashPromotion selectByPrimaryKey(Integer id);
+    SmsFlashPromotion selectByPrimaryKey(Long id);
 
     int updateByExampleSelective(@Param("record") SmsFlashPromotion record, @Param("example") SmsFlashPromotionExample example);
